+++ scripts/build_shapes.py
@@ -222,19 +222,6 @@ def get_GADM_layer(
         # in the GADM processing of sub-national zones
         geodf_temp["GADM_ID"] = geodf_temp[f"GID_{cur_layer_id}"]
 
-        # from pypsa-earth-sec
-        # if layer_id == 0:
-        #     geodf_temp["GADM_ID"] = geodf_temp[f"GID_{cur_layer_id}"].apply(
-        #         lambda x: two_2_three_digits_country(x[:2])
-        #     ) + pd.Series(range(1, geodf_temp.shape[0] + 1)).astype(str)
-        # else:
-        #     # create a subindex column that is useful
-        #     # in the GADM processing of sub-national zones
-        #     # Fix issues with missing "." in selected cases
-        #     geodf_temp["GADM_ID"] = geodf_temp[f"GID_{cur_layer_id}"].apply(
-        #         lambda x: x if x[3] == "." else x[:3] + "." + x[3:]
-        #     )
-
         # append geodataframes
         geodf_list.append(geodf_temp)
 
@@ -1338,6 +1325,14 @@ def gadm(
         lambda x: x if x.find(".") == 0 else "." + x
     )
     df_gadm.set_index("GADM_ID", inplace=True)
+    if df_gadm.index.duplicated().any():
+        all_duplicated = df_gadm.index.duplicated(keep=False).sum()
+        list_duplicated = df_gadm.index[df_gadm.index.duplicated(keep=False)].unique()
+        n_duplicated = df_gadm.index.duplicated(keep="first").sum()
+        logger.warning(
+            f"Duplicated GADM_ID found, dissolving {all_duplicated} geometries with the same GADM_ID into {n_duplicated} shapes: {list_duplicated}"
+        )
+        df_gadm = df_gadm.dissolve(by=df_gadm.index)
 
     if simplify_gadm:
         df_gadm["geometry"] = df_gadm["geometry"].map(